Add `triangles`
I think the tests could be improved.  Some of the NetworkX tests get coverage,
but only compare to 0 triangles.  Also, we should test more with self-edges.

There may be better ways to compute triangles for:
- all nodes
- a subset of nodes
- a single node

There are *a lot* of different ways to compute triangles, so this could be
explored further in the future.  I hope the current PR is competitive.